Init Task3

+++ src/main/java/com/example/task3/Cache.java
@@ -8,5 +8,5 @@ import java.lang.annotation.Target;
 @Target(ElementType.METHOD)
 @Retention(RetentionPolicy.RUNTIME)
 public @interface Cache {
-        int value() default 0;
+    long value() default 0L;
 }
